requirements.txt
torch==2.1.0
torchvision==0.16.0
timm==0.9.12
opencv-python==4.8.0.76
mediapipe==0.10.9
numpy==1.24.4
pillow==10.0.1
pandas==2.0.3